separate game logic from framework
